requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-classic

openai
python-dotenv

pypdf
faiss-cpu
sentence-transformers
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz

hdbscan
umap-learn

keybert
rouge-score
pandas
numpy

scikit-learn